Create simple Jenkins config client.
